+++ parser/options.cc
@@ -114,6 +114,7 @@ void Options::PublishMembers()
   // executable options first
   publish("inputFileName",         &Options::inputFileName);
   publish("visMacroFileName",      &Options::visMacroFileName);
+  publish("geant4MacroFileName",   &Options::geant4MacroFileName);
   publish("visDebug",              &Options::visDebug);
   publish("outputFileName",        &Options::outputFileName);
   publish("outputFormat",          &Options::outputFormat);